+++ packages/najm-next/test-dom/appClient-lifetime.test.tsx
@@ -106,10 +106,10 @@ describe("NajmAppProvider mount lifetime", () => {
 
     render(
       <AppRouterContext.Provider value={router}>
-        <NajmAppProvider authClient={authClient} snapshot={makeSnapshot("disabled", 1)}>
+        <NajmAppProvider authClient={authClient} query={true} snapshot={makeSnapshot("disabled", 1)}>
           <ClientProbe />
         </NajmAppProvider>
-        <NajmAppProvider authClient={authClient} snapshot={makeSnapshot("disabled", 2)}>
+        <NajmAppProvider authClient={authClient} query={true} snapshot={makeSnapshot("disabled", 2)}>
           <ClientProbe />
         </NajmAppProvider>
       </AppRouterContext.Provider>,